Move set_inp/out_map to DeeplayModule
